Stock and special size logic also use case creation
