--- capitole-visibility-application/src/main/java/com/capitole/visibility/agregate/Stock.java
@@ -1,9 +1,0 @@
-package com.capitole.visibility.agregate;
-
-import com.capitole.visibility.vo.SizeId;
-import com.capitole.visibility.vo.StockQuantity;
-
-public record Stock(SizeId id,
-                    StockQuantity stockQuantity) {
-
-}
