--- capitole-visibility-application/src/main/java/com/capitole/visibility/entity/Size.java
@@ -1,13 +1,0 @@
-package com.capitole.visibility.entity;
-
-import com.capitole.visibility.agregate.Stock;
-import com.capitole.visibility.vo.SizeBackSoon;
-import com.capitole.visibility.vo.SizeId;
-import com.capitole.visibility.vo.SizeSpecial;
-
-public record Size(SizeId id,
-                   SizeBackSoon sizeBackSoon,
-                   SizeSpecial sizeSpecial,
-                   Stock stock) {
-
-}
