+++ capitole-visibility-application/src/main/java/com/capitole/visibility/handler/GetProductToShowUseCase.java
@@ -1,0 +1,9 @@
+package com.capitole.visibility.handler;
+
+import java.util.logging.Logger;
+
+public class GetProductToShowUseCase {
+
+    Logger LOGGER = Logger.getLogger(GetProductToShowUseCase.class.getName());
+
+}
