--- capitole-visibility-application/src/main/java/com/capitole/visibility/vo/ProductId.java
@@ -1,4 +1,0 @@
-package com.capitole.visibility.vo;
-
-public record ProductId(int value) {
-}
