--- capitole-visibility-application/src/main/java/com/capitole/visibility/vo/SizeId.java
@@ -1,4 +1,0 @@
-package com.capitole.visibility.vo;
-
-public record SizeId(int value) {
-}
