--- capitole-visibility-application/src/main/java/com/capitole/visibility/vo/SizeSpecial.java
@@ -1,4 +1,0 @@
-package com.capitole.visibility.vo;
-
-public record SizeSpecial(boolean value) {
-}
