+++ capitole-visibility-domain/src/main/java/com/capitole/visibility/domain/entity/Size.java
@@ -1,0 +1,17 @@
+package com.capitole.visibility.domain.entity;
+
+import com.capitole.visibility.domain.agregate.Stock;
+import com.capitole.visibility.domain.vo.SizeBackSoon;
+import com.capitole.visibility.domain.vo.SizeId;
+import com.capitole.visibility.domain.vo.SizeSpecial;
+
+public record Size(SizeId id,
+                   SizeBackSoon sizeBackSoon,
+                   SizeSpecial sizeSpecial,
+                   Stock stock) {
+
+    public boolean inStock(){
+        return sizeBackSoon().value() || stock().inStock();
+    }
+
+}
