+++ capitole-visibility-domain/src/main/java/com/capitole/visibility/domain/ports/SizeRepository.java
@@ -1,0 +1,4 @@
+package com.capitole.visibility.domain.ports;
+
+public interface SizeRepository {
+}
