+++ capitole-visibility-domain/src/main/java/com/capitole/visibility/domain/vo/ProductId.java
@@ -1,0 +1,4 @@
+package com.capitole.visibility.domain.vo;
+
+public record ProductId(int value) {
+}
