+++ capitole-visibility-domain/src/main/java/com/capitole/visibility/domain/vo/ProductSequence.java
@@ -1,4 +1,4 @@
-package com.capitole.visibility.vo;
+package com.capitole.visibility.domain.vo;
 
 
 public record ProductSequence(int value) {
