+++ capitole-visibility-domain/src/main/java/com/capitole/visibility/domain/vo/SizeId.java
@@ -1,0 +1,4 @@
+package com.capitole.visibility.domain.vo;
+
+public record SizeId(int value) {
+}
