+++ capitole-visibility-domain/src/main/java/com/capitole/visibility/domain/vo/SizeSpecial.java
@@ -1,4 +1,4 @@
-package org.visibility.domain.vo;
+package com.capitole.visibility.domain.vo;
 
 public record SizeSpecial(boolean value) {
 }
