--- capitole-visibility-domain/src/main/java/org/visibility/domain/agregate/Stock.java
@@ -1,9 +1,0 @@
-package org.visibility.domain.agregate;
-
-import com.capitole.visibility.vo.SizeId;
-import com.capitole.visibility.vo.StockQuantity;
-
-public record Stock(SizeId id,
-                    StockQuantity stockQuantity) {
-
-}
