--- capitole-visibility-domain/src/main/java/org/visibility/domain/vo/ProductSequence.java
@@ -1,5 +1,0 @@
-package org.visibility.domain.vo;
-
-
-public record ProductSequence(int value) {
-}
