--- capitole-visibility-domain/src/main/java/org/visibility/domain/vo/SizeBackSoon.java
@@ -1,4 +1,0 @@
-package org.visibility.domain.vo;
-
-public record SizeBackSoon(boolean value) {
-}
