--- capitole-visibility-domain/src/main/java/org/visibility/domain/vo/SizeId.java
@@ -1,4 +1,0 @@
-package org.visibility.domain.vo;
-
-public record SizeId(int value) {
-}
